Breaks /api calls out into seperate route
Now the app.js file isn't the only place for the routes. This
commit breaks out the api calls into a seperate route file.
